refactor: do not export config entry and do some cleanups
In this commit, we avoid re-exporting items from the config entry point to prevent bundling
these libraries together. Additionally, some functions have been removed because they are
only used once. The `OnPush` strategy has also been improved to make it more automatic.
