+++ README.md
@@ -129,7 +129,8 @@ export const tooltipVariation = {
 ### Use `Component` as content
 
 ```ts
-import { injectTippyRef, TippyInstance } from '@ngneat/helipopper/config';
+import type { TippyInstance } from '@ngneat/helipopper/config';
+import { injectTippyRef } from '@ngneat/helipopper';
 
 @Component()
 class MyComponent {
@@ -315,7 +316,8 @@ tpVisible = new EventEmitter<boolean>();
 ### Create `tippy` Programmatically
 
 ```typescript
-import { TippyService, TippyInstance } from '@ngneat/helipopper';
+import type { TippyInstance } from '@ngneat/helipopper/config';
+import { TippyService } from '@ngneat/helipopper';
 
 class Component {
   @ViewChild('inputName') inputName: ElementRef;
