+++ projects/ngneat/helipopper/config/src/public-api.ts
@@ -9,8 +9,6 @@ export {
   TippyConfig,
   TippyLoader,
   TIPPY_LOADER,
-  TIPPY_REF,
   TIPPY_CONFIG,
 } from './tippy.types';
 export { provideTippyLoader, provideTippyConfig } from './providers';
-export { injectTippyRef } from './inject-tippy';
