+++ projects/ngneat/helipopper/src/lib/inject-tippy.ts
@@ -1,6 +1,11 @@
-import { inject } from '@angular/core';
+import { inject, InjectionToken } from '@angular/core';
+import type { TippyInstance } from '@ngneat/helipopper/config';
 
-import { TIPPY_REF, TippyErrorCode, type TippyInstance } from './tippy.types';
+import { TippyErrorCode } from './utils';
+
+export const TIPPY_REF = /* @__PURE__ */ new InjectionToken<TippyInstance>(
+  typeof ngDevMode !== 'undefined' && ngDevMode ? 'TIPPY_REF' : ''
+);
 
 export function injectTippyRef(): TippyInstance {
   const instance = inject(TIPPY_REF, { optional: true });
