+++ projects/ngneat/helipopper/src/public-api.ts
@@ -1,16 +1,4 @@
 export { TippyDirective } from './lib/tippy.directive';
 export { TippyService } from './lib/tippy.service';
 export { inView, overflowChanges } from './lib/utils';
-
-export {
-  ExtendedTippyInstance,
-  TippyInstance,
-  TIPPY_REF,
-  TippyConfig,
-  TIPPY_CONFIG,
-  tooltipVariation,
-  popperVariation,
-  withContextMenuVariation,
-  provideTippyConfig,
-  injectTippyRef,
-} from '@ngneat/helipopper/config';
+export { TIPPY_REF, injectTippyRef } from './lib/inject-tippy';
